Rename file product to products for BO Products page
