+++ tests/puppeteer/campaigns/sanity/productsBO/02_CRUDStandardProductInBO.js
@@ -5,7 +5,7 @@ const {expect} = require('chai');
 const LoginPage = require('../../../pages/BO/login');
 const DashboardPage = require('../../../pages/BO/dashboard');
 const BOBasePage = require('../../../pages/BO/BObasePage');
-const ProductPage = require('../../../pages/BO/product');
+const ProductsPage = require('../../../pages/BO/products');
 const AddProductPage = require('../../../pages/BO/addProduct');
 const FOProductPage = require('../../../pages/FO/product');
 const ProductFaker = require('../../data/faker/product');
@@ -14,7 +14,7 @@ let page;
 let loginPage;
 let dashboardPage;
 let boBasePage;
-let productPage;
+let productsPage;
 let addProductPage;
 let foProductPage;
 let productData;
@@ -26,7 +26,7 @@ const init = async () => {
   loginPage = await (new LoginPage(page));
   dashboardPage = await (new DashboardPage(page));
   boBasePage = await (new BOBasePage(page));
-  productPage = await (new ProductPage(page));
+  productsPage = await (new ProductsPage(page));
   addProductPage = await (new AddProductPage(page));
   foProductPage = await (new FOProductPage(page));
   productData = await (new ProductFaker());
@@ -44,17 +44,18 @@ global.scenario('Create, read, update and delete Standard product in BO', async 
   });
   test('should go to Products page', async () => {
     await boBasePage.goToSubMenu(boBasePage.productsParentLink, boBasePage.productsLink);
-    const pageTitle = await productPage.getPageTitle();
-    await expect(pageTitle).to.contains(productPage.pageTitle);
+    const pageTitle = await productsPage.getPageTitle();
+    await expect(pageTitle).to.contains(productsPage.pageTitle);
   });
   test('should reset all filters', async () => {
-    if (await productPage.elementVisible(productPage.filterResetButton, 2000)) await productPage.resetFilter();
-    await productPage.resetFilterCategory();
-    const numberOfProducts = await productPage.getNumberOfProductsFromList();
+    if (await productsPage.elementVisible(productsPage.filterResetButton, 2000))
+      await productsPage.resetFilter();
+    await productsPage.resetFilterCategory();
+    const numberOfProducts = await productsPage.getNumberOfProductsFromList();
     await expect(numberOfProducts).to.be.above(0);
   });
   test('should create Product', async () => {
-    await productPage.goToAddProductPage();
+    await productsPage.goToAddProductPage();
     const createProductMessage = await addProductPage.createEditProduct(productData);
     await expect(createProductMessage).to.equal('Settings updated.');
   });
@@ -74,8 +75,8 @@ global.scenario('Create, read, update and delete Standard product in BO', async 
   });
   test('should delete Product and be on product list page', async () => {
     const testResult = await addProductPage.deleteProduct();
-    await expect(testResult).to.equal(productPage.productDeletedSuccessfulMessage);
-    const pageTitle = await productPage.getPageTitle();
-    await expect(pageTitle).to.contains(productPage.pageTitle);
+    await expect(testResult).to.equal(productsPage.productDeletedSuccessfulMessage);
+    const pageTitle = await productsPage.getPageTitle();
+    await expect(pageTitle).to.contains(productsPage.pageTitle);
   });
 }, init, true);
