+++ tests/puppeteer/pages/BO/products.js
